Move common Docker functionality to docker_util package
